+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `Array::with_codec_specific_options` and `Array::set_codec_specific_options` to reconfigure codec chains with `CodecSpecificOptions`
+- Add `ShardingCodecOptions` for runtime configuration of the sharding codec
+- Add `CodecChain::with_codec_specific_options`
+
+### Changed
+- Bump `zarrs_codec` to 0.2.1
+
 ## [0.23.6] - 2026-03-08
 
 ### Changed